move event.ts into `@base/event` module
