+++ cocos/core/event/index.ts
@@ -26,3 +26,4 @@
 export { EventTarget } from './event-target';
 export { Eventify } from './eventify';
 export { AsyncDelegate } from './async-delegate';
+export { Event } from './event';
