+++ cocos/input/types/event/event-acceleration.ts
@@ -23,7 +23,7 @@
  THE SOFTWARE.
 */
 
-import { Event } from './event';
+import { Event } from '../../../core/event';
 import { Acceleration } from '../acceleration';
 import { SystemEventType } from '../event-enum';
