+++ cocos/input/types/event/event-gamepad.ts
@@ -29,7 +29,7 @@
  */
 
 import { GamepadInputDevice as Gamepad } from 'pal/input';
-import { Event } from './event';
+import { Event } from '../../../core/event';
 
 /**
   * @en
