+++ cocos/input/types/event/event-handle.ts
@@ -23,7 +23,7 @@
 */
 
 import { HandleInputDevice } from 'pal/input';
-import { Event } from './event';
+import { Event } from '../../../core/event';
 import { SystemEventTypeUnion } from '../event-enum';
 
 /**
