+++ cocos/input/types/event/event-hmd.ts
@@ -23,7 +23,7 @@
 */
 
 import { HMDInputDevice } from 'pal/input';
-import { Event } from './event';
+import { Event } from '../../../core/event';
 import { SystemEventTypeUnion } from '../event-enum';
 
 /**
