+++ cocos/input/types/event/index.ts
@@ -22,7 +22,6 @@
  THE SOFTWARE.
 */
 
-export * from './event';
 export * from './event-acceleration';
 export * from './event-keyboard';
 export * from './event-mouse';
